requirements.txt
asgiref==3.7.2
boto3==1.28.49
botocore==1.31.49
Django==4.2.5
django-restframework==0.0.1
django-storages==1.14
djangorestframework==3.14.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
jmespath==1.0.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
Pillow==10.0.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.6.2
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.16
